requirements.txt
Flask==3.0.3
google-genai==1.3.0
twilio==9.3.2
gunicorn==23.0.0
flask-cors==5.0.0
python-dotenv==1.0.1
